fix(frontend): cast substring() position param to fix bot-hit slug query
Without the (::int) cast, Postgres couldn't infer the parameter's type
and silently picked the regex-match substring(text, pattern) overload
instead of the position-based one -- returned NULL (or a bogus regex
match) for nearly every row instead of the real slug. Broke the Vercel
build: generateStaticParams got a null slug for /artista/[slug].

+++ frontend/lib/db/disco.ts
@@ -115,8 +115,12 @@ export const getDiscoMeta = unstable_cache(
 // traffic (bot_hits) rather than lastfm_listeners, so it prioritizes the pages
 // Googlebot/bots actually hit, independent of a record's popularity elsewhere.
 export async function getTopBotHitSlugs(pathPrefix: string, limit: number): Promise<string[]> {
+  // The (::int) cast is load-bearing: without it, Postgres can't infer the
+  // parameter's type and picks the regex-match `substring(text, pattern)`
+  // overload instead of the position-based one, silently returning NULL (or
+  // a bogus regex match) for almost every row instead of the real slug.
   const rows = await prisma.$queryRaw<{ slug: string }[]>`
-    SELECT substring(path from ${pathPrefix.length + 1}) AS slug, count(*) AS hits
+    SELECT substring(path from (${pathPrefix.length + 1})::int) AS slug, count(*) AS hits
     FROM bot_hits
     WHERE path LIKE ${pathPrefix + "%"}
     GROUP BY slug
